requirements.txt
requests
sounddevice
numpy
faiss-cpu
sentence-transformers>=3.0.1
openai-whisper
termcolor
openwakeword